Extract common signer test functionality
